backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
passlib
argon2-cffi
python-jose
sentence-transformers
chromadb
requests
pydantic
python-dotenv
pymupdf
python-docx
openpyxl
pandas
google-genai